Fix error handling for message deletion

+++ src/commands/devonly/deletemsg.ts
@@ -36,13 +36,20 @@ module.exports = {
         interaction.channel?.messages
             .fetch(messageId)
             .then(async msg => {
+                if (msg.author.id !== process.env.BOT_USER_ID) {
+                    return interaction.reply({
+                        content: `You can only delete messages sent by the bot.`,
+                        flags: MessageFlags.Ephemeral,
+                    });
+                }
+
                 try {
                     msg.delete();
                 } catch (error) {
                     console.error("[ERROR] Failed to delete message: ", error);
 
                     return interaction.reply({
-                        content: `Failed to delete message with ID \`${messageId}\``,
+                        content: `Failed to delete message with ID \`${messageId}\`.`,
                         flags: MessageFlags.Ephemeral,
                     });
                 }
@@ -54,7 +61,6 @@ module.exports = {
             })
             .catch(error => {
                 console.error("[ERROR] Failed to fetch message: ", error);
-
                 interaction.reply({
                     content: `Message with ID \`${messageId}\` not found`,
                     flags: MessageFlags.Ephemeral,
